refactor(frontend): split manage tokens services in respective networks
